+++ backend/app/models/product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float, Text, Integer
+from sqlalchemy import Column, String, Numeric, Text, Integer
 from app.db.database import Base
 
 
@@ -8,7 +8,7 @@ class Product(Base):
     id = Column(String, primary_key=True, index=True)
     name = Column(String, nullable=False, index=True)
     description = Column(Text, nullable=True)
-    price = Column(Float, nullable=True)
+    price = Column(Numeric(10, 2), nullable=True)
     category = Column(String, nullable=True, index=True)
     material = Column(String, nullable=True)
     availability = Column(String, nullable=True)
